add data.js and util.js modules
